Implement recoding to represent travel cases
Recodes certain tips with their suspected division of infection, represented as `division_exposure`. Performs DTA on this trait. By default, the map uses this, but it's coloured by the division of sample collection.

Notes for running: Lat/longs & colours are automatically created from the `division` trait. Values only appearing in the recoded version should be added as a normal division and the snakefile will take care of the rest. 

The collection of tips & their known travel history is incomplete.